move lmsusage
